src/chat-app/requirements.txt
chainlit ~= 1.2.0
langchain-openai ~= 0.2.1
langchain ~= 0.3.1
langchain-community ~= 0.3.1
python-dotenv ~= 1.0.1
azure-search-documents ~= 11.5.1
azure-identity ~= 1.18.0

langgraph ~= 0.2.32
pyodbc ~= 5.1.0